factor runtime meta to meta.scala
